requirements.txt
alchemy==20.5
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2025.1.31
chardet==3.0.4
click==8.1.8
dnspython==2.7.0
docutils==0.21.2
email_validator==2.2.0
exceptiongroup==1.2.2
factory_boy==3.3.3
Faker==37.1.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.0.12
greenlet==3.2.1
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==2.8
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mock==5.2.0
motor==3.7.1
orjson==3.10.16
packaging==25.0
passlib==1.7.4
pluggy==1.5.0
pydantic==2.11.3
pydantic-extra-types==2.10.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.2
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-mock==3.14.0
python-daemon==2.2.4
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.22.0
rich==14.0.0
rich-toolkit==0.14.3
shellingham==1.5.4
sniffio==1.3.1
sql==2022.4.0
SQLAlchemy==2.0.40
starlette==0.46.2
tomli==2.2.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==1.25.11
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
